+++ CHANGELOG.md
@@ -1,3 +1,7 @@
+### Unreleased
+- Fix: `NoMethodError: undefined method `position' for nil:NilClass` in /gems/pronto-0.10.0/lib/pronto/formatter/github_pull_request_review_formatter.rb:20:in `line_number'.
+- [2](https://github.com/pdobb/pronto-bundler_audit/issues/2) Fix: `NoMethodError: undefined method `repo' for #<Pronto::BundlerAudit::GemfileLock::Scanner::Patch:...>`.
+
 ### 0.4.0 - 2019-05-08
 - Remove patch-level processing... just always scan Gemfile.lock when this runner is invoked.
 
@@ -15,7 +19,6 @@
 ### 0.1.1 - 2019-04-29
 - Add line number to Pronto::Message; fixes GitHub API usage error when attempting to add errors to PR comments
 - Add gem version requirements to gemspec
-
 
 ### 0.1.0 - 2019-04-28
 - Initial release!